# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePortfolio API",
        "description": "API para el sistema de portfolio electrónico de estudiantes",
        "contact": {
            "name": "ePortfolio Support",
            "email": "admin@eportfolio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "http://eportfolio.test/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://eportfolio.cifpcarlos3.es/api/v1",
            "description": "Production server"
        },
        {
            "url": "https://eportfolio.juezlti.eu/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/users/{parent_id}/modulos-formativos": {
            "get": {
                "tags": [
                    "ModuloFormativo"
                ],
                "summary": "List all moduloformativos",
                "description": "Retrieve a paginated list of moduloformativos",
                "operationId": "a36b17b92170f5cbcce48364b6402714",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ModuloFormativo"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ModuloFormativo"
                ],
                "summary": "Create a new moduloformativo",
                "description": "Create a new moduloformativo resource",
                "operationId": "6cb46288912d62a04d25eb3d91dfccd9",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreModuloFormativoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ModuloFormativo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{parent_id}/modulos-formativos/{id}": {
            "get": {
                "tags": [
                    "ModuloFormativo"
                ],
                "summary": "Show a specific moduloformativo",
                "description": "Retrieve a specific moduloformativo by ID",
                "operationId": "57f6f1d779825fec758f4e6a233faa96",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the moduloformativo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ModuloFormativo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ModuloFormativo"
                ],
                "summary": "Update a specific moduloformativo",
                "description": "Update a specific moduloformativo by ID",
                "operationId": "65b7fc301d314c28ad718d9a85a1c550",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the moduloformativo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateModuloFormativoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ModuloFormativo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ModuloFormativo"
                ],
                "summary": "Delete a specific moduloformativo",
                "description": "Delete a specific moduloformativo by ID",
                "operationId": "6e4c47e200ca958fb2327cb21c76cbba",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the ciclo formativo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "ModuloFormativo eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreModuloFormativoRequest": {
                "title": "Store Módulo Formativo Request",
                "description": "Datos requeridos para crear un Módulo Formativo",
                "required": [
                    "nombre",
                    "codigo",
                    "horas_totales",
                    "curso_escolar",
                    "centro"
                ],
                "properties": {
                    "ciclo_formativo_id": {
                        "description": "ID del ciclo formativo",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del módulo formativo",
                        "type": "string",
                        "maxLength": 255
                    },
                    "codigo": {
                        "description": "Código del módulo formativo",
                        "type": "string",
                        "maxLength": 50
                    },
                    "horas_totales": {
                        "description": "Horas totales del módulo",
                        "type": "integer",
                        "minimum": 1
                    },
                    "curso_escolar": {
                        "description": "Curso escolar (ej: 2024-2025)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "centro": {
                        "description": "Centro educativo",
                        "type": "string",
                        "maxLength": 255
                    },
                    "docente_id": {
                        "description": "ID del docente",
                        "type": "integer"
                    },
                    "descripcion": {
                        "description": "Descripción del módulo",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateModuloFormativoRequest": {
                "title": "Update Módulo Formativo Request",
                "description": "Datos para actualizar un Módulo Formativo",
                "properties": {
                    "ciclo_formativo_id": {
                        "description": "ID del ciclo formativo",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del módulo formativo",
                        "type": "string",
                        "maxLength": 255
                    },
                    "codigo": {
                        "description": "Código del módulo formativo",
                        "type": "string",
                        "maxLength": 50
                    },
                    "horas_totales": {
                        "description": "Horas totales del módulo",
                        "type": "integer",
                        "minimum": 1
                    },
                    "curso_escolar": {
                        "description": "Curso escolar (ej: 2024-2025)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "centro": {
                        "description": "Centro educativo",
                        "type": "string",
                        "maxLength": 255
                    },
                    "docente_id": {
                        "description": "ID del docente",
                        "type": "integer"
                    },
                    "descripcion": {
                        "description": "Descripción del módulo",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ModuloFormativo": {
                "title": "Módulo Formativo",
                "description": "Modelo de Módulo Formativo",
                "properties": {
                    "id": {
                        "description": "ID único",
                        "type": "integer"
                    },
                    "ciclo_formativo_id": {
                        "description": "ID del ciclo formativo",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del módulo formativo",
                        "type": "string"
                    },
                    "codigo": {
                        "description": "Código del módulo formativo",
                        "type": "string"
                    },
                    "horas_totales": {
                        "description": "Horas totales del módulo",
                        "type": "integer"
                    },
                    "curso_escolar": {
                        "description": "Curso escolar",
                        "type": "string"
                    },
                    "centro": {
                        "description": "Centro educativo",
                        "type": "string"
                    },
                    "docente_id": {
                        "description": "ID del docente",
                        "type": "integer"
                    },
                    "descripcion": {
                        "description": "Descripción del módulo",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}